safer-ws/requirements.txt
pymongo==4.3.3
flask-marshmallow==0.15.0
flasgger==0.9.5
apispec==6.3.0
flask==2.2.4
python-dotenv==1.0.0
marshmallow-mongoengine==0.31.2
flask-cors==3.0.10
openrouteservice
geopandas
requests
locust